Move generateAlias back to apply-query
