+++ api/src/database/run-ast.ts
@@ -6,9 +6,8 @@ import { clone, cloneDeep, isNil, merge, pick, uniq } from 'lodash-es';
 import { PayloadService } from '../services/payload.js';
 import type { AST, FieldNode, FunctionFieldNode, M2ONode, NestedCollectionNode } from '../types/ast.js';
 import { applyFunctionToColumnName } from '../utils/apply-function-to-column-name.js';
-import type { ColumnSortRecord } from '../utils/apply-query.js';
+import { type ColumnSortRecord, generateAlias } from '../utils/apply-query.js';
 import applyQuery, { applyLimit, applySort } from '../utils/apply-query.js';
-import { generateAlias } from '../utils/generate-alias.js';
 import { getCollectionFromAlias } from '../utils/get-collection-from-alias.js';
 import type { AliasMap } from '../utils/get-column-path.js';
 import { getColumn } from '../utils/get-column.js';
