--- api/src/utils/generate-alias.ts
@@ -1,3 +1,0 @@
-import { customAlphabet } from 'nanoid/non-secure';
-
-export const generateAlias = customAlphabet('abcdefghijklmnopqrstuvwxyz', 5);
